feature/29 content drag and drop component
