+++ src/App.tsx
@@ -6,9 +6,9 @@ import Home from './pages/Home';
 import Content from './pages/Content';
 import Style from './pages/Style';
 
-import Tpo from './Pages/Tpo';
-import Footer from '@components/footer';
-import Detail from './Pages/Detail';
+import Tpo from './pages/Tpo';
+//import Footer from '@components/pages/Footer';
+import Detail from './pages/Detail';
 import Sign from './pages/Sign';
 import SignUp from './components/SignUp';
 import ContentDetail from 'pages/ContentDetail';
